+++ src/components/project/TileView.svelte
@@ -473,9 +473,11 @@
             var(--wordplay-border-color);
     }
 
-    .tile:not(:global(.output)).responsive,
-    .tile:not(:global(.output)).horizontal {
+    .tile.responsive,
+    .tile.horizontal {
         border-right: var(--wordplay-border-width) solid
+            var(--wordplay-border-color);
+        border-bottom: var(--wordplay-border-width) solid
             var(--wordplay-border-color);
     }